src/maintenance/requirements.txt
tinytag==1.7.0
PyYAML==5.4.1
SQLAlchemy==1.4.14